requirements-airflow.txt
mysql-connector-python==8.3.0
pandas==2.2.1
